phone/main_test.go
package main

import "testing"

type testCase struct {
	input string
	want  string
}

func TestNormalize(t *testing.T) {
	testCases := []testCase{
		{"123 456 7891", "1234567891"},
		{"(123) 456 7892", "1234567892"},
		{"(123) 456-7893", "1234567893"},
		{"123-456-7894", "1234567894"},
		{"123-456-7890", "1234567890"},
		{"1234567892", "1234567892"},
		{"(123)456-7892", "1234567892"},
	}
	for _, tc := range testCases {
		t.Run(tc.input, func(t *testing.T) {
			actual := normalize(tc.input)
			if actual != tc.want {
				t.Errorf("want %s; got %s", tc.want, actual)
			}
		})
	}
}
